Refactor: MOD change datasource with yaml setting
- properties.yaml 설정으로 datasource 를 변경할 수 있도록 개선
